Version FInal
Subo la version que es para entrega, ya que al cambiar de pc me confundi
de version

+++ operacionesMat.h
@@ -2,4 +2,4 @@ float Sumar(float,float);
 float Restar(float,float);
 float Dividir(float,float);
 float Multiplicar(float,float);
-int Factorial(float);
+float Factorial(float);
